Fix for trying to delete item when it's the only item shown. Fix for trying to delete item when there are no items selected.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C15649

+++ FormItemEditor.cs
@@ -261,9 +261,12 @@ namespace Rawr
 		private void buttonDelete_Click(object sender, EventArgs e)
 		{
 			_changingItemCache = true;
-			ItemCache.DeleteItem(listViewItems.SelectedItems[0].Tag as Item);
-			listViewItems.Items.Remove(listViewItems.SelectedItems[0]);
-			listViewItems.SelectedIndices.Add(0);
+            if (listViewItems.SelectedItems.Count > 0)
+            {
+                ItemCache.DeleteItem(listViewItems.SelectedItems[0].Tag as Item);
+                listViewItems.Items.Remove(listViewItems.SelectedItems[0]);
+                if (listViewItems.Items.Count > 0) listViewItems.SelectedIndices.Add(0);
+            }
 			_changingItemCache = false;
 		}
